+++ configure.ac
@@ -5047,10 +5047,12 @@ CARGO_BUILD="$CARGO build"
 CARGO_CLEAN="$CARGO clean"
 CARGO_TEST="$CARGO test"
 CARGO_CLIPPY="$CARGO clippy"
+CARGO_RUN="$CARGO run"
 
 AC_SUBST(CARGO_BUILD)
 AC_SUBST(CARGO_CLEAN)
 AC_SUBST(CARGO_CLIPPY)
+AC_SUBST(CARGO_RUN)
 
 AC_ARG_ENABLE(rust-debug,
 [AS_HELP_STRING([--enable-rust-debug],
