+++ rust_src/.gitignore
@@ -1,1 +1,2 @@
 target
+generated/
