+++ rust_src/remacs-bindings/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "remacs-bindings"
+version = "0.1.0"
+description = "Tool to create remacs C bindings."
+edition = "2018"
+
+[dependencies]
+libc = "0.2"
+regex = "0.2"
+bindgen = { git = "https://github.com/rust-lang-nursery/rust-bindgen.git", rev = "badb49277dddf1ea5d407075f9deea48897b52df" }
